Make `Model` a default root namespace for command and handler classes.
