+++ tests/test_app/Command/FooBusCommand.php
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-namespace App\Bus\Command;
+namespace App\Command;
 
 /**
  * @author Robert Pustułka <r.pustulka@robotusers.com>
